testing cuda + openGL

+++ include/constants.h
@@ -2,7 +2,7 @@
 #define CONSTANTS_H
 
 namespace constants{
-    extern int particleCount;
+    extern size_t particleCount;
     extern float gridSizeX;
     extern float gridSizeY;
     extern float gridSizeZ;
